Refactor files into packages
