+++ data/rainy.txt
@@ -1,3 +1,3 @@
-T | 0 | meet wiwiwi
 T | 1 | meerakjgn
-E | 0 | project meeting | 2019-12-02T18:00 | 2019-12-02T18:00
+E | 1 | project meeting | 2019-12-02T18:00 | 2019-12-02T18:00
+T | 0 | kwjenf
