+++ src/main/java/commands/ByeCommand.java
@@ -1,3 +1,10 @@
+package commands;
+
+import exception.RainyException;
+import storage.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
