+++ src/main/java/exception/RainyException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class RainyException extends Exception {
     public RainyException(String message) {
         super(message);
