+++ src/main/java/tasks/Event.java
@@ -1,3 +1,7 @@
+package tasks;
+
+import exception.RainyException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
