+++ src/main/java/tasks/Task.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
