+++ src/main/java/tasks/Todo.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
