+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Rainy
+Hello! I'm rainy.Rainy
 What can I do for you?
 ____________________________________________________________
